+++ cmd/template/cluster/flag.go
@@ -17,10 +17,22 @@ const (
 	flagProvider = "provider"
 
 	// AWS only.
-	flagExternalSNAT       = "external-snat"
-	flagEKS                = "aws-eks"
-	flagPodsCIDR           = "pods-cidr"
-	flagControlPlaneSubnet = "control-plane-subnet"
+	flagAWSExternalSNAT       = "external-snat"
+	flagAWSEKS                = "aws-eks"
+	flagAWSControlPlaneSubnet = "control-plane-subnet"
+
+	// OpenStack only.
+	flagOpenStackCloud                = "cloud"
+	flagOpenStackCloudConfig          = "cloud-config"
+	flagOpenStackDNSNameservers       = "dns-nameservers"
+	flagOpenStackExternalNetworkID    = "external-network-id"
+	flagOpenStackFailureDomain        = "failure-domain"
+	flagOpenStackImageName            = "image-name"
+	flagOpenStackNodeMachineFlavor    = "node-machine-flavor"
+	flagOpenStackNodeCIDR             = "node-cidr"
+	flagOpenStackRootVolumeDiskSize   = "root-volume-disk-size"
+	flagOpenStackRootVolumeSourceType = "root-volume-source-type"
+	flagOpenStackRootVolumeSourceUUID = "root-volume-source-uuid"
 
 	// Common.
 	flagClusterIDDeprecated = "cluster-id"
@@ -31,18 +43,36 @@ const (
 	flagOutput              = "output"
 	flagOrganization        = "organization"
 	flagOwner               = "owner" // TODO: Remove some time after December 2021
+	flagPodsCIDR            = "pods-cidr"
 	flagRelease             = "release"
 	flagLabel               = "label"
 )
 
-type flag struct {
-	Provider string
-
-	// AWS only.
+type awsFlag struct {
 	ControlPlaneSubnet string
 	ExternalSNAT       bool
 	EKS                bool
-	PodsCIDR           string
+}
+
+type openStackFlag struct {
+	Cloud                string   // OPENSTACK_CLOUD
+	CloudConfig          string   // <no equivalent env var>
+	DNSNameservers       []string // OPENSTACK_DNS_NAMESERVERS
+	ExternalNetworkID    string   // <no equivalent env var>
+	FailureDomain        string   // OPENSTACK_FAILURE_DOMAIN
+	ImageName            string   // OPENSTACK_IMAGE_NAME
+	NodeMachineFlavor    string   // OPENSTACK_NODE_MACHINE_FLAVOR
+	RootVolumeDiskSize   string   // <no equivalent env var>
+	RootVolumeSourceType string   // <no equivalent env var>
+	RootVolumeSourceUUID string   // <no equivalent env var>
+	NodeCIDR             string   // <no equivalent env var>
+}
+
+type flag struct {
+	Provider string
+
+	AWS       awsFlag
+	OpenStack openStackFlag
 
 	// Common.
 	ClusterIDDeprecated string
@@ -53,6 +83,7 @@ type flag struct {
 	Output              string
 	Organization        string
 	Owner               string
+	PodsCIDR            string
 	Release             string
 	Label               []string
 
@@ -64,10 +95,35 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Provider, flagProvider, "", "Installation infrastructure provider.")
 
 	// AWS only.
-	cmd.Flags().StringVar(&f.ControlPlaneSubnet, flagControlPlaneSubnet, "", "Subnet used for the Control Plane.")
-	cmd.Flags().BoolVar(&f.ExternalSNAT, flagExternalSNAT, false, "AWS CNI configuration.")
-	cmd.Flags().BoolVar(&f.EKS, flagEKS, false, "Enable EKS. Only available for AWS Release v20.0.0 (CAPA)")
-	cmd.Flags().StringVar(&f.PodsCIDR, flagPodsCIDR, "", "CIDR used for the pods.")
+	cmd.Flags().StringVar(&f.AWS.ControlPlaneSubnet, flagAWSControlPlaneSubnet, "", "Subnet used for the Control Plane.")
+	cmd.Flags().BoolVar(&f.AWS.ExternalSNAT, flagAWSExternalSNAT, false, "AWS CNI configuration.")
+	cmd.Flags().BoolVar(&f.AWS.EKS, flagAWSEKS, false, "Enable AWSEKS. Only available for AWS Release v20.0.0 (CAPA)")
+
+	// OpenStack only.
+	cmd.Flags().StringVar(&f.OpenStack.Cloud, flagOpenStackCloud, "openstack", "Name of cloud (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.CloudConfig, flagOpenStackCloudConfig, "cloud-config", "Name of cloud config (OpenStack only).")
+	cmd.Flags().StringSliceVar(&f.OpenStack.DNSNameservers, flagOpenStackDNSNameservers, nil, "DNS nameservers (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.ExternalNetworkID, flagOpenStackExternalNetworkID, "", "External network ID (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.FailureDomain, flagOpenStackFailureDomain, "", "Failure domain (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.ImageName, flagOpenStackImageName, "ubuntu-2004-kube-v1.20.9", "Image name (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.NodeMachineFlavor, flagOpenStackNodeMachineFlavor, "", "Node machine flavor (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.RootVolumeDiskSize, flagOpenStackRootVolumeDiskSize, "", "Root volume disk size (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.RootVolumeSourceType, flagOpenStackRootVolumeSourceType, "", "Root volume source type (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.RootVolumeSourceUUID, flagOpenStackRootVolumeSourceUUID, "", "Root volume source UUID (OpenStack only).")
+	cmd.Flags().StringVar(&f.OpenStack.NodeCIDR, flagOpenStackNodeCIDR, "", "CIDR used for the nodes.")
+
+	// TODO: Make these flags visible once we have a better method for displaying provider-specific flags.
+	_ = cmd.Flags().MarkHidden(flagOpenStackCloud)
+	_ = cmd.Flags().MarkHidden(flagOpenStackCloudConfig)
+	_ = cmd.Flags().MarkHidden(flagOpenStackDNSNameservers)
+	_ = cmd.Flags().MarkHidden(flagOpenStackExternalNetworkID)
+	_ = cmd.Flags().MarkHidden(flagOpenStackFailureDomain)
+	_ = cmd.Flags().MarkHidden(flagOpenStackImageName)
+	_ = cmd.Flags().MarkHidden(flagOpenStackNodeMachineFlavor)
+	_ = cmd.Flags().MarkHidden(flagOpenStackRootVolumeDiskSize)
+	_ = cmd.Flags().MarkHidden(flagOpenStackRootVolumeSourceType)
+	_ = cmd.Flags().MarkHidden(flagOpenStackRootVolumeSourceUUID)
+	_ = cmd.Flags().MarkHidden(flagOpenStackNodeCIDR)
 
 	// Common.
 	cmd.Flags().StringVar(&f.ClusterIDDeprecated, flagClusterIDDeprecated, "", "Unique identifier of the cluster (deprecated).")
@@ -78,11 +134,12 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Output, flagOutput, "", "File path for storing CRs.")
 	cmd.Flags().StringVar(&f.Organization, flagOrganization, "", "Workload cluster organization.")
 	cmd.Flags().StringVar(&f.Owner, flagOwner, "", "Workload cluster owner organization (deprecated).")
+	cmd.Flags().StringVar(&f.PodsCIDR, flagPodsCIDR, "", "CIDR used for the pods.")
 	cmd.Flags().StringVar(&f.Release, flagRelease, "", "Workload cluster release.")
 	cmd.Flags().StringSliceVar(&f.Label, flagLabel, nil, "Workload cluster label.")
 
 	// TODO: Make this flag visible when we roll CAPA/EKS out for customers
-	_ = cmd.Flags().MarkHidden(flagEKS)
+	_ = cmd.Flags().MarkHidden(flagAWSEKS)
 
 	// TODO: Remove the flag completely some time after August 2021
 	_ = cmd.Flags().MarkDeprecated(flagMasterAZ, "please use --control-plane-az.")
@@ -124,7 +181,8 @@ func (f *flag) Validate() error {
 	validProviders := []string{
 		key.ProviderAWS,
 		key.ProviderAzure,
-		key.ProviderVsphere,
+		key.ProviderOpenStack,
+		key.ProviderVSphere,
 	}
 	isValidProvider := false
 	for _, p := range validProviders {
@@ -159,10 +217,10 @@ func (f *flag) Validate() error {
 			return microerror.Maskf(invalidFlagError, "--%s must only contain alphanumeric characters, and start with a letter", flagName)
 		}
 
-		if f.ControlPlaneSubnet != "" {
-			matchedSubnet, err := regexp.MatchString("^20|21|22|23|24|25$", f.ControlPlaneSubnet)
+		if f.AWS.ControlPlaneSubnet != "" {
+			matchedSubnet, err := regexp.MatchString("^20|21|22|23|24|25$", f.AWS.ControlPlaneSubnet)
 			if err == nil && !matchedSubnet {
-				return microerror.Maskf(invalidFlagError, "--%s must be a valid subnet size (20, 21, 22, 23, 24 or 25)", flagControlPlaneSubnet)
+				return microerror.Maskf(invalidFlagError, "--%s must be a valid subnet size (20, 21, 22, 23, 24 or 25)", flagAWSControlPlaneSubnet)
 			}
 		}
 	}
@@ -197,6 +255,12 @@ func (f *flag) Validate() error {
 		case key.ProviderAzure:
 			if len(f.ControlPlaneAZ) > 1 {
 				return microerror.Maskf(invalidFlagError, "--%s supports one availability zone only", flagControlPlaneAZ)
+			}
+		case key.ProviderOpenStack:
+			if f.OpenStack.NodeCIDR != "" {
+				if !validateCIDR(f.OpenStack.NodeCIDR) {
+					return microerror.Maskf(invalidFlagError, "--%s must be a valid CIDR", flagOpenStackNodeCIDR)
+				}
 			}
 		}
 	}