+++ internal/key/provider.go
@@ -1,8 +1,9 @@
 package key
 
 const (
-	ProviderAWS     = "aws"
-	ProviderAzure   = "azure"
-	ProviderKVM     = "kvm"
-	ProviderVsphere = "vsphere"
+	ProviderAWS       = "aws"
+	ProviderAzure     = "azure"
+	ProviderKVM       = "kvm"
+	ProviderOpenStack = "openstack"
+	ProviderVSphere   = "vsphere"
 )